# docs/openapi.yaml
openapi: 3.0.3
info:
  title: rust-ticket-seckill API
  version: 0.1.0
  description: |
    Minimal ticket seckill system (planned contract).
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string }
        activity_id: { type: string, format: uuid }
        ticket_type_id: { type: string, format: uuid }
        qty: { type: integer }
        amount_cents: { type: integer }
        status: { type: string, enum: [pending, paid, cancelled, expired] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...